# api/openapi.yaml
openapi: 3.0.3
info:
  title: MoonLightCloud API
  description: >-
    Device lifecycle + versioned sync ingestion with the first Sale-sync
    vertical slice: durable inbox, sale.finalized.v1 validated before ACK,
    async projection with durable processing state. Business APIs are
    versioned under /api/v1. Health and version stay outside the version
    boundary. ACK means durable PostgreSQL commit, never queueing alone.
  version: 2D
paths:
...
  /api/v1/dashboard/auth/login:
    post:
      summary: Primary (password) authentication; starts an MFA stage
      description: >-
        Same-origin only. Never yields a FULL session directly: the response
        stage is MFA_PENDING (verify a TOTP or recovery code) or MFA_SETUP
        (enroll TOTP first). Every failure (unknown account, wrong password,
        disabled account) is the same 401; account and network throttling
        answer 429 for a bounded time.
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthLoginRequest'}
      responses:
        '200':
          description: Password accepted; challenge session cookie set (HttpOnly, SameSite=Strict, Secure + __Host- prefix outside development).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResult'}
        '400':
          description: Malformed or oversized body.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '401':
          description: INVALID_CREDENTIALS (generic, identical for every cause).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: CSRF_REJECTED (missing or foreign Origin).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '429':
          description: RATE_LIMITED (temporary account or network throttle); Retry-After set.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/dashboard/auth/activate:
    post:
      summary: Consume a one-time activation token and set the first password
      description: >-
        Single-use, short-lived (24 h) token created by an OWNER. Returns an
        MFA_SETUP session; the account becomes ACTIVE only after TOTP
        enrollment.
      operationId: authActivate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token, password]
              properties:
                token: {type: string, maxLength: 128}
                password: {type: string, description: '12 to 256 characters; never truncated.'}
      responses:
        '200':
          description: Activated; MFA_SETUP session cookie set.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResult'}
        '400':
          description: ACTIVATION_INVALID (unknown, used or expired) or WEAK_PASSWORD.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: CSRF_REJECTED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/dashboard/auth/me:
    get:
      summary: Current human, stage, permissions and Store access
      description: Works in every session stage. Returns no credential material other than the session CSRF token.
      operationId: authMe
      security:
        - dashboardSession: []
      responses:
        '200':
          description: Current session.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthMe'}
        '401':
          description: AUTH_REQUIRED (no, unknown, expired, revoked or superseded session).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/auth/logout:
    post:
      summary: Revoke the server-side session and clear the cookie
      operationId: authLogout
      security:
        - dashboardSession: []
      responses:
        '200':
          description: Session revoked; replaying the old cookie is rejected.
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated: {type: boolean}
        '401':
          description: AUTH_REQUIRED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: CSRF_REJECTED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/dashboard/auth/mfa/verify:
    post:
      summary: Complete an MFA_PENDING session with a TOTP or one-time recovery code
      description: >-
        RFC 6238 TOTP (30 s, 6 digits, one adjacent step tolerated); a step is
        accepted once (replay rejected). Recovery codes are single-use. On
        success the session is rotated to FULL (old cookie revoked).
      operationId: authMFAVerify
      security:
        - dashboardSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                code: {type: string, pattern: '^[0-9 ]{6,7}$'}
                recovery_code: {type: string, maxLength: 32}
      responses:
        '200':
          description: FULL session; rotated cookie set.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResult'}
        '401':
          description: INVALID_MFA_CODE or AUTH_REQUIRED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: CSRF_REJECTED or the session is not MFA_PENDING.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '429':
          description: RATE_LIMITED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/dashboard/auth/mfa/enroll/start:
    post:
      summary: Generate the TOTP secret for an MFA_SETUP session (shown once)
      description: The secret is stored only AES-256-GCM encrypted under AUTH_MFA_ENCRYPTION_KEY.
      operationId: authMFAEnrollStart
      security:
        - dashboardSession: []
      responses:
        '200':
          description: Provisioning material.
          content:
            application/json:
              schema:
                type: object
                required: [secret, otpauth_uri]
                properties:
                  secret: {type: string, description: 'Base32 secret for manual authenticator entry.'}
                  otpauth_uri: {type: string}
        '401':
          description: AUTH_REQUIRED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: CSRF_REJECTED or the session is not MFA_SETUP.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/dashboard/auth/mfa/enroll/confirm:
    post:
      summary: Verify the first TOTP code, enable MFA and receive recovery codes once
      operationId: authMFAEnrollConfirm
      security:
        - dashboardSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code: {type: string}
      responses:
        '200':
          description: MFA enabled; FULL session; ten one-time recovery codes (stored only as digests).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResult'}
        '401':
          description: INVALID_MFA_CODE or AUTH_REQUIRED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: CSRF_REJECTED or the session is not MFA_SETUP.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/dashboard/auth/mfa/recovery-codes:
    post:
      summary: Regenerate recovery codes (previous codes stop working)
      operationId: authRecoveryCodes
      security:
        - dashboardSession: []
      responses:
        '200':
          description: New codes, shown once.
          content:
            application/json:
              schema:
                type: object
                properties:
                  recovery_codes: {type: array, items: {type: string}}
        '401':
          description: AUTH_REQUIRED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: CSRF_REJECTED or MFA not completed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/dashboard/auth/password:
    post:
      summary: Change own password (other sessions revoked, current rotated)
      operationId: authChangePassword
      security:
        - dashboardSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password, new_password]
              properties:
                current_password: {type: string}
                new_password: {type: string}
      responses:
        '200':
          description: Changed; rotated session.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResult'}
        '400':
          description: WEAK_PASSWORD.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '401':
          description: INVALID_CREDENTIALS (current password) or AUTH_REQUIRED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: CSRF_REJECTED or MFA not completed.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /api/v1/dashboard/users/{id}/{action}:
    post:
      summary: Account administration (users.manage; mfa-reset needs security.manage)
      description: >-
        action: activation (new one-time token, PENDING_SETUP only) | role
        {role} | status {status: ACTIVE|DISABLED} | memberships {all_stores,
        store_ids} | mfa-reset | sessions-revoke. Role/status/MFA changes
        invalidate the target's sessions. The last active OWNER can never be
        disabled or demoted (409 LAST_OWNER), enforced transactionally.
      operationId: usersAction
      security:
        - dashboardSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: action
          in: path
          required: true
          schema: {type: string, enum: [activation, role, status, memberships, mfa-reset, sessions-revoke]}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                role: {type: string, enum: [OWNER, ADMIN]}
                status: {type: string, enum: [ACTIVE, DISABLED]}
                all_stores: {type: boolean}
                store_ids: {type: array, items: {type: string, format: uuid}}
      responses:
        '200':
          description: Updated account, activation token, or revocation result.
          content:
            application/json:
              schema: {type: object}
        '400':
          description: INVALID_INPUT.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '401':
          description: AUTH_REQUIRED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: FORBIDDEN (permission, Store reach) or CSRF_REJECTED.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '404':
          description: Unknown account or action.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '409':
          description: LAST_OWNER or STATE_CONFLICT.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
components:
  securitySchemes:
    deviceAuth:
      type: http
      scheme: bearer
      bearerFormat: '<device-id>.<credential-id>.<secret-hex> (raw secret shown once at provisioning/rotation)'
      description: >-
        Authorization: Bearer <device_uuid>.<credential_uuid>.<64-hex secret>.
        Never in query parameters. HTTPS in production.
    reportAuth:
      type: http
      scheme: bearer
      bearerFormat: 'REPORTING_API_TOKEN (temporary report-read secret until dashboard user auth)'
      description: >-
        Authorization: Bearer <reporting-api-token>. Report scope only;
        device sync credentials are never valid here. Fail-closed: required
        16+ chars in staging/production; open reports need ALL of
        ENVIRONMENT=development, ALLOW_UNAUTHENTICATED_REPORTING=true, and no
        token. Anything else (including missing ENVIRONMENT) fails startup.
    dashboardSession:
      type: apiKey
      in: cookie
      name: __Host-mlc_session
      description: >-
        Opaque server-side human session (ADR-0053): 256-bit random token,
        stored only as a digest. Cookie __Host-mlc_session (HttpOnly, Secure,
        SameSite=Strict, Path=/) outside development; mlc_session over plain
        HTTP in development. Every dashboard route also requires the FULL
        (MFA-complete) stage, a named permission and server-side Store
        membership; mutations additionally require a same-origin Origin/Referer
        and the X-CSRF-Token header. Device credentials and
        REPORTING_API_TOKEN are never valid here.
...
  schemas:
...
    AuthLoginRequest:
      type: object
      additionalProperties: false
      required: [login, password]
      properties:
        login: {type: string, maxLength: 254, description: 'Email or username; trimmed and lower-cased.'}
        password: {type: string}
...